python_bot/requirements.txt
discord.py==2.3.2
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
APScheduler>=3.10.0
openpyxl>=3.0.0
gspread>=6.0.0
google-auth>=2.0.0
gtts>=2.5.0
pydub>=0.25.0
PyNaCl>=1.5.0
gunicorn>=21.0.0
